+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.3-wip
 
-- Added a validation step on the output of the build and link hooks.
+- Added a validation step on the output of the build and link hooks (both as a
+  per package, and as in all the packages together).
 
 ## 0.8.2